+++ packages/components/ng-ovh-otrs/src/otrs-popup/otrs-popup.constant.js
@@ -96,6 +96,11 @@ export const API_EXCLUDED = {
     '/ip/loadbalancing',
     '/license/office',
     '/license/virtuozzo',
+    '/dedicated/nvmeof',
+    '/license/hycu',
+    '/license/officePrepaid',
+    '/products',
+    '/internalMonitoring',
   ],
 };