+++ packages/manager/apps/container/src/container/nav-reshuffle/sidebar/navigation-tree/services/hostedPrivateCloud.ts
@@ -1,6 +1,9 @@
+import illustration from '@/assets/images/sidebar/hosted-private-cloud.png';
 export default {
   id: 'hosted-private-cloud',
   translation: 'sidebar_hpc',
+  shortTranslation: 'sidebar_hpc_short',
+  illustration,
   features: ['hosted-private-cloud'],
   children: [
     {