+++ packages/manager/apps/key-management-service/src/api/services/index.ts
@@ -1,0 +1,3 @@
+export * from './get';
+export * from './put';
+export * from '../hooks/useUpdateOkmsName';