+++ packages/manager/apps/key-management-service/src/hooks/guide/useGuideUtils.tsx
@@ -1,0 +1,46 @@
+import { useContext, useEffect, useState } from 'react';
+import { CountryCode } from '@ovh-ux/manager-config';
+import { ShellContext } from '@ovh-ux/manager-react-shell-client';
+import { GUIDE_LIST } from './guidesLinks.constant';
+
+const helpCenterUrl = 'https://help.ovhcloud.com';
+
+type GetGuideLinkProps = {
+  name?: string;
+  subsidiary: CountryCode | string;
+};
+
+function getGuideListLink({ subsidiary }: GetGuideLinkProps) {
+  const list: { [guideName: string]: string } = {};
+  const keys = Object.entries(GUIDE_LIST);
+  keys.forEach((key) => {
+    const guide =
+      GUIDE_LIST[key[0]][subsidiary as CountryCode] || GUIDE_LIST[key[0]].GB;
+
+    list[key[0]] = helpCenterUrl + guide;
+  });
+  return list;
+}
+
+interface GuideLinkProps {
+  [guideName: string]: string;
+}
+
+function useGuideUtils() {
+  const { shell } = useContext(ShellContext);
+  const { environment } = shell;
+  const [list, setList] = useState({});
+
+  useEffect(() => {
+    const getSubSidiary = async () => {
+      const env = await environment.getEnvironment();
+      const { ovhSubsidiary } = env.getUser();
+      const guideList = getGuideListLink({ subsidiary: ovhSubsidiary });
+      setList(guideList);
+    };
+    getSubSidiary();
+  }, []);
+  return list as GuideLinkProps;
+}
+
+export default useGuideUtils;