+++ packages/manager/modules/log-to-customer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.1](https://github.com/ovh/manager/compare/@ovh-ux/manager-log-to-customer@1.1.0...@ovh-ux/manager-log-to-customer@1.1.1) (2024-06-25)
+
+
+### Bug Fixes
+
+* fix log to customer superseded css for oui-datagrid ([#12057](https://github.com/ovh/manager/issues/12057)) ([f80289c](https://github.com/ovh/manager/commit/f80289cba6ea9a0257439a896f4d90bae3ad6474))
+
+
+
+
+
 # [1.1.0](https://github.com/ovh/manager/compare/@ovh-ux/manager-log-to-customer@1.0.0...@ovh-ux/manager-log-to-customer@1.1.0) (2024-04-24)